ch: rename folders
