+++ src/pages/index.js
@@ -2,8 +2,8 @@ import Home from './Home/Home';
 // import DeiBadge from './DeiBadge/DeiBadge';
 
 import Projects from './Projects/Projects';
-import About from './About/About';
+import About from './AboutDeiBadging/About';
 import BadgingRequirements from './BadgingRequirements/BadgingRequirements';
-import Page from './ApplyForBadge/SelectProjectRepo';
+import Page from './SelectProjectForBadging/SelectProjectRepo';
 
 export { Home, About, Projects, BadgingRequirements, Page };
